refactors code into more distinct units so it's easier to find and maintain certain functionality
